Changed a lot of constants, check description
Made the force on the tower more snappy - I realized it is quite difficult to keep the archers on

Increased the size of the tower spoke hitboxes - the archers won't fall out by sliding as often

Decreased the minimum distance before the archer is registered as lost

Added a force down on the ball to counteract the force that goes upwards on the tower (can be adjusted)

Added a minimum velocity of the ball, so the camera shakes only when the ball hits the ground fast (can be adjusted)

Added an experimental script to set the water x and z pos to that of the player. This has an advantage that we don't have to tile the water objects, it just moves with the player. Unfortunately it just looks wierd, and might actually induce sea-sickness, lol